# perf(http/file_server): Avoid calculating Content-Type when 304 Not Modified response