close #3318 

If the file server returns a 304 Not Modified, calls `Deno.open()` and `contentType()` are unnecessary. Change those functions to be called on a later line.

This change seems to make the 304 response generation slightly (100µs) faster. (446µs -> 340µs)

before:
```
benchmark                      time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------------- -----------------------------
fileServer_200OK           551.74 µs/iter(388.4 µs … 1.58 ms) 621.5 µs 1.01 ms 1.24 ms
fileServer_304NotModified  446.54 µs/iter(350.7 µs … 1.23 ms) 486.3 µs 883.2 µs 996.6 µs
```

after:
```
benchmark                      time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------------- -----------------------------
fileServer_200OK           536.2 µs/iter(364.9 µs … 51.06 ms) 542.5 µs 1.02 ms 1.2 ms
fileServer_304NotModified  340.07 µs/iter(257.8 µs … 2.58 ms) 365 µs 653.9 µs 695.1 µs
```

<details>
<summary>
benchmark code:
</summary>

```ts
import { serveFile } from "./file_server.ts";

{
  const req = new Request("http://localhost/");
  Deno.bench(async function fileServer_200OK() {
    await serveFile(req, "./README.md");
  });
}
{
  const etag =
    (await serveFile(new Request("http://localhost/"), "./README.md"))
      .headers.get("etag")!;

  const req = new Request("http://localhost/", {
    headers: { "if-none-match": etag },
  });
  Deno.bench(async function fileServer_304NotModified() {
    await serveFile(req, "./README.md");
  });
}
```

</details>